buyingLunch file
It's like a game. Some random people are sitting in a restaurant and one of them has to order food for all. But they can't decide who is that person. So they wrote a program in which a function will be there and it takes name of all the persons. and a random person's name will be generated and that person has to buy food.